refactor (redis): move internal/redis to internal/redis/redis
