+++ internal/redis/redis/aggregator_worker.go
